requirements.txt
einops==0.8.0
local-attention==1.9.14
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
patool==1.12
reformer-pytorch==1.4.4
scikit-learn==1.2.2
scipy==1.10.1
sktime==0.16.1
sympy==1.11.1
#torch==1.7.1
tqdm==4.64.1
openpyxl==3.1.5
pyarrow==17.0.0
calflops==0.3.2
transformers==4.46.3